Address review follow-ups for renumber and wrap
definitions.rs: spell "finalized" with the Oxford -ize ending so the
comment matches both the en-GB-oxendict convention used elsewhere in
the project and the existing `finalize_numeric_candidates` function
name.

tests/wrap_unit.rs: split the 430-line integration test into focused
submodules under `tests/wrap_unit/` so each file sits within the
400-line repository limit. The entry point now wires the
`code_spans`, `prefixed`, `footnotes`, `shell_blocks`, and `stream`
submodules with `#[path]`. Each submodule carries its own imports;
the `assert_footnote_reference_is_intact` helper moves into the
footnotes submodule (its sole caller). Insta snapshot files keep
their reviewed content; they are renamed without the `wrap_unit__`
prefix to match the path insta now picks for the relocated tests,
with `prepend_module_to_snapshot = false` pointing them back to
`tests/snapshots/`.
